requirements.txt
streamlit==1.44.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.1
plotly==6.0.1
Pillow==11.1.0
ultralytics==8.3.111
torch==2.6.0
torchvision==0.21.0
seaborn
pyyaml